chore(server): move prependEntries util back where it was
