+++ lib/utils/addEntries.js
@@ -2,7 +2,6 @@
 
 const webpack = require('webpack');
 const createDomain = require('./createDomain');
-const prependEntries = require('./prependEntries');
 
 /**
  * A Entry, it can be of type string or string[] or Object<string | string[],string>
@@ -54,6 +53,42 @@ function addEntries(originalCompiler, options, server) {
   } else if (options.hot) {
     hotEntry = require.resolve('webpack/hot/dev-server');
   }
+
+  /**
+   * prependEntries Method
+   * @param {Entry} originalEntry
+   * @param {Entry} additionalEntries
+   * @returns {Entry}
+   */
+  const prependEntries = (originalEntry, additionalEntries) => {
+    if (typeof originalEntry === 'function') {
+      return () =>
+        Promise.resolve(originalEntry()).then((entry) =>
+          prependEntries(entry, additionalEntries)
+        );
+    }
+
+    if (typeof originalEntry === 'object' && !Array.isArray(originalEntry)) {
+      const entryObj = {};
+
+      Object.keys(originalEntry).forEach((key) => {
+        const entryDescription = originalEntry[key];
+        entryObj[key] = prependEntries(entryDescription, additionalEntries);
+      });
+
+      return entryObj;
+    }
+
+    // in this case, originalEntry is a string or an array.
+    // make sure that we do not add duplicates.
+    const entriesClone = additionalEntries.slice(0);
+    [].concat(originalEntry).forEach((newEntry) => {
+      if (!entriesClone.includes(newEntry)) {
+        entriesClone.push(newEntry);
+      }
+    });
+    return entriesClone;
+  };
 
   /**
    *
